Add files via upload
Coded a tic-tac-toe game in Java with a partner. Uses a Binary Search Tree to store all possible combinations of a TicTacToe game. When playing against the computer, the program traverses the BST in order to come up with the next best possible move, with the highest probability of winning. The TicTacToe game is displayed to the console using Java GUI, complete with sound effects, and thread delays to create a more seamless playing experience.